Rename tcp to opaque
