+++ linkerd/app/outbound/src/endpoint.rs
@@ -263,7 +263,7 @@ impl<S> Outbound<S> {
         // external discovery.
         let http = self
             .clone()
-            .push_tcp_endpoint::<http::Connect>()
+            .push_opaque_endpoint::<http::Connect>()
             .push_http_endpoint()
             .map_stack(|config, rt, stk| {
                 stk.push_on_service(
@@ -281,7 +281,7 @@ impl<S> Outbound<S> {
             .push_http_server()
             .into_inner();
 
-        let opaque = self.push_tcp_endpoint().push_tcp_forward();
+        let opaque = self.push_opaque_endpoint().push_opaque_forward();
 
         opaque.push_detect_http(http).map_stack(|_, _, stk| {
             stk.instrument(|e: &tcp::Endpoint| info_span!("forward", endpoint = %e.addr))
