+++ linkerd/app/outbound/src/opaque.rs
@@ -1,0 +1,4 @@
+pub mod concrete;
+pub mod endpoint;
+pub mod forward;
+pub mod logical;
