+++ linkerd/app/outbound/src/tcp.rs
@@ -6,9 +6,7 @@ use linkerd_app_core::{
     Error,
 };
 
-pub mod concrete;
 pub mod connect;
-pub mod logical;
 pub mod opaque_transport;
 
 pub use self::connect::Connect;
